pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "nest-asyncio>=1.6.0",
    "numpy>=2.2.5",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "scipy>=1.15.2",
    "telethon>=1.40.0",
]
